# Add checked arithmetic for `/`