This also changes the implementation of `/` for `Fixed`. The new `/` also checks for overflow.

`/` (`checked_fdiv`) throws `DivideError` for division by zero and `OverflowError` for overflow.
```julia
julia> 0.5Q0f7 / 0Q0f7 # v0.8.4 and this PR
ERROR: DivideError: integer division error

julia> 0.5N0f8 / 0N0f8 # v0.8.4
ERROR: ArgumentError: Normed{UInt8,8} is an 8-bit type representing 256 values from 0.0 to 1.0; cannot represent Inf

julia> 0.5N0f8 / 0N0f8 # this PR
ERROR: DivideError: integer division error
```
```julia
julia> -1Q0f7 / -1Q0f7 # v0.8.4
-1.0Q0f7

julia> -1Q0f7 / -1Q0f7 # this PR
ERROR: OverflowError: -1.0Q0f7 / -1.0Q0f7 overflowed for type Q0f7

julia> 1N0f8 / 0.5N0f8 # v0.8.4
ERROR: ArgumentError: Normed{UInt8,8} is an 8-bit type representing 256 values from 0.0 to 1.0; cannot represent 1.9921874

julia> 1N0f8 / 0.5N0f8 # this PR
ERROR: OverflowError: 1.0N0f8 / 0.502N0f8 overflowed for type N0f8
```