@kimikage 2020-08-29
## Benchmark
```julia
x_q0f7  = rand(Q0f7,  1000, 1000); y_q0f7  = rand(Q0f7,  1000, 1000);
x_q0f15 = rand(Q0f15, 1000, 1000); y_q0f15 = rand(Q0f15, 1000, 1000);
x_q3f4  = rand(Q3f4,  1000, 1000); y_q3f4  = rand(Q3f4,  1000, 1000);
x_q3f12 = rand(Q3f12, 1000, 1000); y_q3f12 = rand(Q3f12, 1000, 1000);

x_n0f8  = rand(N0f8,  1000, 1000); y_n0f8  = rand(N0f8,  1000, 1000);
x_n0f16 = rand(N0f16, 1000, 1000); y_n0f16 = rand(N0f16, 1000, 1000);
x_n4f4  = rand(N4f4,  1000, 1000); y_n4f4  = rand(N4f4,  1000, 1000);
x_n4f12 = rand(N4f12, 1000, 1000); y_n4f12 = rand(N4f12, 1000, 1000);

non_zero(x) = x === zero(x) ? eps(x) : x;
z(x::X) where X = clamp(x * 0.5 * rand(X), X);

w_q0f7  = non_zero.(y_q0f7);  z_q0f7  = z.(w_q0f7);
w_q0f15 = non_zero.(y_q0f15); z_q0f15 = z.(w_q0f15);
w_q3f4  = non_zero.(y_q3f4);  z_q3f4  = z.(w_q3f4);
w_q3f12 = non_zero.(y_q3f12); z_q3f12 = z.(w_q3f12);

w_n0f8  = non_zero.(y_n0f8);  z_n0f8  = z.(w_n0f8);
w_n0f16 = non_zero.(y_n0f16); z_n0f16 = z.(w_n0f16);
w_n4f4  = non_zero.(y_n4f4);  z_n4f4  = z.(w_n4f4);
w_n4f12 = non_zero.(y_n4f12); z_n4f12 = z.(w_n4f12);

@btime $z_q0f7  ./ $w_q0f7;
@btime $z_q0f15 ./ $w_q0f15;
@btime $z_q3f4  ./ $w_q3f4;
@btime $z_q3f12 ./ $w_q3f12;

@btime $z_n0f8  ./ $w_n0f8;
@btime $z_n0f16 ./ $w_n0f16;
@btime $z_n4f4  ./ $w_n4f4;
@btime $z_n4f12 ./ $w_n4f12;
```
```julia
@btime wrapping_fdiv.($x_q0f7 , $y_q0f7 );
@btime wrapping_fdiv.($x_q0f15, $y_q0f15);
@btime wrapping_fdiv.($x_q3f4 , $y_q3f4 );
@btime wrapping_fdiv.($x_q3f12, $y_q3f12);

@btime wrapping_fdiv.($x_n0f8 , $y_n0f8 );
@btime wrapping_fdiv.($x_n0f16, $y_n0f16);
@btime wrapping_fdiv.($x_n4f4 , $y_n4f4 );
@btime wrapping_fdiv.($x_n4f12, $y_n4f12);

@btime saturating_fdiv.($x_q0f7 , $y_q0f7 );
@btime saturating_fdiv.($x_q0f15, $y_q0f15);
@btime saturating_fdiv.($x_q3f4 , $y_q3f4 );
@btime saturating_fdiv.($x_q3f12, $y_q3f12);

@btime saturating_fdiv.($x_n0f8 , $y_n0f8 );
@btime saturating_fdiv.($x_n0f16, $y_n0f16);
@btime saturating_fdiv.($x_n4f4 , $y_n4f4 );
@btime saturating_fdiv.($x_n4f12, $y_n4f12);
```


| `Fixed`            | before | after  | | `Normed`           | before | after  |
|:------------------:|-------:|-------:|-|:------------------:|-------:|-------:|
|`z_q0f7  ./ w_q0f7` |2.016 ms|1.880 ms| |`z_n0f8  ./ w_n0f8` |2.157 ms|1.656 ms|
|`z_q0f15 ./ w_q0f15`|2.435 ms|2.576 ms| |`z_n0f16 ./ w_n0f16`|3.017 ms|2.393 ms|
|`z_q3f4  ./ w_q3f4` |2.016 ms|1.940 ms| |`z_n4f4  ./ w_n4f4` |2.216 ms|1.691 ms|
|`z_q3f12 ./ w_q3f12`|2.458 ms|2.599 ms| |`z_n4f12 ./ w_n4f12`|3.611 ms|2.362 ms|

|                  |`Fixed`             |        | | `Normed`           |        |
|-----------------:|:------------------:|-------:|-|:------------------:|-------:|
|  `wrapping_fidv.`|`(x_q0f7 , y_q0f7 )`|1.452 ms| |`(x_n0f8 , y_n0f8 )`|1.409 ms|
|  `wrapping_fidv.`|`(x_q0f15, y_q0f15)`|2.099 ms| |`(x_n0f16, y_n0f16)`|2.140 ms|
|  `wrapping_fidv.`|`(x_q3f4 , y_q3f4 )`|1.456 ms| |`(x_n4f4 , y_n4f4 )`|1.411 ms|
|  `wrapping_fidv.`|`(x_q3f12, y_q3f12)`|2.093 ms| |`(x_n4f12, y_n4f12)`|2.078 ms|
|`saturating_fidv.`|`(x_q0f7 , y_q0f7 )`|272.3 μs| |`(x_n0f8 , y_n0f8 )`|217.4 μs|
|`saturating_fidv.`|`(x_q0f15, y_q0f15)`|892.7 μs| |`(x_n0f16, y_n0f16)`|826.2 μs|
|`saturating_fidv.`|`(x_q3f4 , y_q3f4 )`|270.5 μs| |`(x_n4f4 , y_n4f4 )`|218.0 μs|
|`saturating_fidv.`|`(x_q3f12, y_q3f12)`|894.2 μs| |`(x_n4f12, y_n4f12)`|834.9 μs|