@kimikage 2020-08-30
JuliaLang/julia #37085 was merged. ("Normed{UInt128,100}"-->"Normed{UInt128, 100}")
So, shall we update the test first? (cf. https://github.com/JuliaMath/FixedPointNumbers.jl/issues/139#issuecomment-683363989)