@kimikage 2020-08-31
FYI, the `Normed` multiplication has also been checked so far. (cf. #213)
```julia
julia> 10N4f12 * 10N4f12 # v0.8.4
ERROR: ArgumentError: Normed{UInt16,12} is a 16-bit type representing 65536 values from 0.0 to 16.0037; cannot represent 100.0

julia> 10N4f12 * 10N4f12 # PR #213 (checked arithmetic for `Fixed`)
ERROR: OverflowError: 10.0N4f12 * 10.0N4f12 overflowed for type N4f12

julia> 10Q4f11 * 10Q4f11 # PR #220 (wrapping arithmetic for `Fixed`)
4.0Q4f11
```
I will soon be asking for feedback on the default arithmetic in v0.9 and the state of arithmetic in v0.10 and later in the Discourse.