@timholy 2020-08-31
Yeah, I think with multiplication I always convert to `float` first.