@kimikage 2020-08-31
I ran `rebase` and moved the 100% tests to "test/common.jl".